Move Spree::Core::ClassConstantizer::Set to its own file
