+++ core/spec/lib/spree/core/class_constantizer/set_spec.rb
